moved ktor_koin to server/koin use TypeSafe route for Events
