+++ src/main/kotlin/org/developerden/codosseum/server/Routes.kt
@@ -8,13 +8,20 @@ import org.developerden.codosseum.model.ChallengeDifficulty
 class Validate {
 
 	@Resource("/{title}")
-	class Title(val parent: Validate = Validate(), @KtorFieldDescription("Title of the challenge") val title: String)
+  class Title(
+    val parent: Validate = Validate(),
+    @KtorFieldDescription("Title of the challenge")
+    val title: String,
+  )
 }
 
 @Resource("/random")
 class RandomChallenge(
-	@KtorFieldDescription("Filter for challenge difficulties. A list of difficulties to _include_ in the search. If empty, all difficulties will be included.") val difficultyFilters: List<ChallengeDifficulty> = emptyList(),
-	@KtorFieldDescription("Filter for challenge tags. A list of tags to _include_ in the search. If empty, all tags will be included.") val tagFilters: List<String> = emptyList()
-) {
+  @KtorFieldDescription("Filter for challenge difficulties. A list of difficulties to _include_ in the search. If empty, all difficulties will be included.")
+  val difficultyFilters: List<ChallengeDifficulty> = emptyList(),
+  @KtorFieldDescription("Filter for challenge tags. A list of tags to _include_ in the search. If empty, all tags will be included.")
+  val tagFilters: List<String> = emptyList(),
+)
 
-}
+@Resource("/events")
+class Events
