+++ src/main/kotlin/org/developerden/codosseum/server/routes/ValidateRoute.kt
@@ -9,8 +9,8 @@ import io.ktor.server.resources.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import org.developerden.codosseum.ChallengesService
-import org.developerden.codosseum.ktor_koin.inject
 import org.developerden.codosseum.server.Validate
+import org.developerden.codosseum.server.koin.inject
 import org.developerden.codosseum.validation.SolutionValidationService
 import org.developerden.codosseum.validation.ValidationResult
